refactor: resultのschemaの場所を変更
